+++ src/extension.ts
@@ -1,5 +1,7 @@
 import * as vscode from 'vscode';
 import { apktool } from './tools';
+import { outputChannel } from './common';
+import { updateTools } from './downloader';
 
 
 export function activate(context: vscode.ExtensionContext) {
@@ -8,23 +10,33 @@ export function activate(context: vscode.ExtensionContext) {
 	// command for opening an apk file for decoding
 	let openApkFileCommand = vscode.commands.registerCommand('apklab.openApkFile', async () => {
 
-		let result = await vscode.window.showOpenDialog({
-			canSelectFolders: false,
-			filters: {
-				APK: ["apk"]
-			},
-			openLabel: "Select an APK file",
+		updateTools().then(async () => {
+			// browse for an APK file
+			let result = await vscode.window.showOpenDialog({
+				canSelectFolders: false,
+				filters: {
+					APK: ["apk"]
+				},
+				openLabel: "Select an APK file",
+			});
+			if (result && result.length === 1) {
+				// disassemble the selected APK file
+				apktool.decodeAPK(result[0].fsPath);
+			} else {
+				vscode.window.showWarningMessage("APKLAB: no apk was file chosen");
+			}
+		}).catch(() => {
+			outputChannel.appendLine("Can't download/update dependencies!");
 		});
-		if (result && result.length === 1) {
-			apktool.decodeAPK(result[0].fsPath);
-		} else {
-			vscode.window.showWarningMessage("APKLAB: no apk was file chosen");
-		}
 	});
 
 	let rebuildAPkFileCommand = vscode.commands.registerCommand("apklab.rebuildApkFile", (uri: vscode.Uri) => {
-		// rebuild apk using Apktool
-		apktool.rebuildAPK(uri.fsPath);
+		updateTools().then(() => {
+			// rebuild apk using Apktool
+			apktool.rebuildAPK(uri.fsPath);
+		}).catch(() => {
+			outputChannel.appendLine("Can't download/update dependencies!");
+		});
 	});
 
 	context.subscriptions.push(openApkFileCommand, rebuildAPkFileCommand);